# Improve performance with optimised Ruy matrix packing.