--- a/larq_compute_engine/core/bgemm_impl_ruy.h
+++ b/larq_compute_engine/core/bgemm_impl_ruy.h
@@ -10,6 +10,7 @@
 #include "ruy/profiler/instrumentation.h"
 #include "tensorflow/lite/kernels/cpu_backend_context.h"
 #include "tensorflow/lite/kernels/cpu_backend_gemm_params.h"
+#include "tensorflow/lite/kernels/cpu_backend_gemm_ruy.h"
 
 using namespace tflite;
 using namespace tflite::cpu_backend_gemm;
@@ -38,17 +39,16 @@ struct BGemmImplUsingRuy {
 
     // Set up the matrix layouts and mul_params.
     ruy::Matrix<TBitpacked> lhs;
-    ruy::MakeSimpleLayout(lhs_params.rows, lhs_params.cols,
-                          ruy::Order::kRowMajor, lhs.mutable_layout());
     ruy::Matrix<TBitpacked> rhs;
-    ruy::MakeSimpleLayout(rhs_params.rows, rhs_params.cols,
-                          ruy::Order::kColMajor, rhs.mutable_layout());
     ruy::Matrix<DstScalar> dst;
-    ruy::MakeSimpleLayout(dst_params.rows, dst_params.cols,
-                          ruy::Order::kColMajor, dst.mutable_layout());
-    lhs.set_data(lhs_data);
-    rhs.set_data(rhs_data);
-    dst.set_data(dst_data);
+    // We allow these matrices to be cached. Note that this doesn't force them
+    // to be cached; it means that the `cache_policy` of the MatrixParams will
+    // be respected.
+    cpu_backend_gemm::detail::MakeRuyMatrix(lhs_params, lhs_data, &lhs,
+                                            /*use_caching=*/true);
+    cpu_backend_gemm::detail::MakeRuyMatrix(rhs_params, rhs_data, &rhs,
+                                            /*use_caching=*/true);
+    cpu_backend_gemm::detail::MakeRuyMatrix(dst_params, dst_data, &dst);
 
     // We have to make this a `const` matrix because otherwise gcc will try to
     // use the non-const versions of `matrix.data()`
@@ -63,13 +63,8 @@ struct BGemmImplUsingRuy {
 
     constexpr auto BGemmCompiledPaths = ruy::kAllPaths;
 
-    // avoid the reference path for production code
     ruy::Path bgemm_runtime_path = ruy_ctx->SelectPath(ruy::kAllPaths);
 
-    // fallback to standard cpp kernel for all architectures that are not
-    // supported yet.
-    // TODO: this needs to be modified as soon as architecture-specific
-    // optimized kernels are added.
 #if RUY_PLATFORM_ARM
     if (bgemm_runtime_path == ruy::Path::kNeonDotprod)
       bgemm_runtime_path = ruy::Path::kNeon;
@@ -79,12 +74,12 @@ struct BGemmImplUsingRuy {
       bgemm_runtime_path = ruy::Path::kStandardCpp;
 #endif
 
-    // For writing bitpacked output, fallback to the standard C++ kernel.
+    // For now, writing bitpacked output only has a C++ implementation.
     if (std::is_same<DstScalar, TBitpacked>::value) {
       bgemm_runtime_path = ruy::Path::kStandardCpp;
     }
 
-    // For now, int8 only has a C++ implementation.
+    // For now, int8 output only has a C++ implementation.
     if (std::is_same<DstScalar, std::int8_t>::value) {
       bgemm_runtime_path = ruy::Path::kStandardCpp;
     }