## What do these changes do?

This PR introduces an optimised Ruy 'packing' routine for 4x4 column-major kernel layouts. This is the kernel layout that is used for the RHS (input activations) matrix in all our optimised assembly kernels, and the Ruy packing for this layout currently falls back to slow reference code. The new packing routine therefore boosts Aarch64 and Arm32 performance.

This PR also enables caching of the packed LHS (weights) matrix. This is done by enabling the `use_caching` flag within our op, and setting the Ruy cache-policy to 'always cache' for the LHS matrix. Note that the RHS matrix would not benefit from caching because it is not static and changes every inference.

This PR should be viewed (and reviewed) in conjunction with #461, upon which it is re-based. Only the final commit is relevant.

## How Has This Been Tested?

CI.

## Benchmark Results

I have benchmarked and this PR and the baseline on my Android phone (a OnePlus 6) and a Raspberry Pi 4B. I used the LCE benchmark utility with `num_runs=250` and report the average and the standard-deviation.

It's important to note that the baseline used here is current master, prior to #461 (specifically, this commit: 30e9b02eceb).

#### OnePlus 6 (1 thread)

| Model         |    Baseline   |       PR      | Change |
|---------------|:-------------:|:-------------:|:------:|
| QuickNet      | 16.45 +- 0.05 | 14.58 +- 0.06 | -11.4% |
| QuickNetLarge | 24.03 +- 0.08 | 21.73 +- 0.08 | -9.6%  |
| QuickNetXL    | 42.03 +- 0.10 | 37.82 +- 0.08 | -10.0% |

#### OnePlus 6 (4 threads)

| Model         |    Baseline   |       PR      | Change |
|---------------|:-------------:|:-------------:|:------:|
| QuickNet      | 6.91 +- 0.23  | 6.39 +- 0.31  | -7.5%  |
| QuickNetLarge | 10.05 +- 0.38 | 9.47 +- 0.33  | -5.8%  |
| QuickNetXL    | 17.06 +- 0.38 | 16.05 +- 0.39 | -5.9%  |

#### Raspberry Pi 4B (1 thread)

| Model         |    Baseline   |       PR      | Change |
|---------------|:-------------:|:-------------:|:------:|
| QuickNet      | 31.68 +- 0.05 | 30.45 +- 0.05 | -3.9%  |
| QuickNetLarge | 48.25 +- 0.49 | 46.49 +- 0.08 | -3.6%  |
| QuickNetXL    | 83.76 +- 0.12 | 80.47 +- 0.19 | -3.9%  |

#### Raspberry Pi 4B (4 threads)

| Model         |    Baseline   |       PR      | Change |
|---------------|:-------------:|:-------------:|:------:|
| QuickNet      | 15.45 +- 0.05 | 14.94 +- 0.07 | -3.3%  |
| QuickNetLarge | 23.00 +- 0.07 | 22.20 +- 0.06 | -3.5%  |
| QuickNetXL    | 36.92 +- 0.11 | 35.41 +- 0.08 | -4.1%  |

There is a strong performance improvement. It is more pronounced on the OnePlus 6 than the Raspberry Pi, but significant on both. It's more than I was expecting!

For reference, the speed-up seems to be fairly evenly split between the new optimised packing routine and the LHS caching.

## Related issue number

N/A.