@AdamHillier 2020-08-13
Note that the baseline benchmark figures I have here for the Raspberry Pi are already a lot better than those reported on our website. I'm not sure why that is, it *might* be because the previous results - used on the website - might have been taken from a benchmark build with the Ruy profiler enabled.