@AdamHillier 2020-08-18
> Did you verify that having the global cache option set to false, but the parameter to `MakeRuyMatrix` set to true, actually caches the matrix? (e.g. by benchmarking Quicknet for example)

Yep, it works nicely :+1: