@AdamHillier 2020-08-24
Note that the benchmark numbers for the Raspberry Pi recorded here (both the baseline and the PR) are slightly questionable - we've had challenges reproducing them. See https://github.com/larq/compute-engine/pull/470#issuecomment-679322167.